refactor(authority-source-files):
change enum name

Closes: MODELINKS-170
