+++ src/test/java/org/folio/entlinks/service/dataloader/SourceFileDeserializerTest.java
@@ -1,7 +1,7 @@
 package org.folio.entlinks.service.dataloader;
 
-import static org.folio.entlinks.domain.entity.AuthoritySourceType.FOLIO;
-import static org.folio.entlinks.domain.entity.AuthoritySourceType.LOCAL;
+import static org.folio.entlinks.domain.entity.AuthoritySourceFileSource.FOLIO;
+import static org.folio.entlinks.domain.entity.AuthoritySourceFileSource.LOCAL;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
